requirements.txt
docker>=6.0.0
pyyaml>=6.0
click>=8.1.3
pytest>=7.0.0
pytest-cov>=4.0.0
distro>=1.8.0
psutil>=5.9.0
typing_extensions>=4.5.0
colorama>=0.4.6
schema>=0.7.5
python-dotenv>=1.0.0

# Notification System
# Note: smtplib is part of the Python standard library, no need for a wrapper
slack-sdk>=3.19.0  # For Slack notifications
discord-webhook>=1.0.0  # For Discord notifications
aiohttp>=3.8.0  # For webhook notifications
html2text>=2020.1.16  # For converting HTML to text
email-validator>=2.0.0  # For validating email addresses

# AI Provider Integration
requests>=2.28.0
tiktoken>=0.3.0
anthropic>=0.5.0
google-generativeai>=0.3.0
openai>=1.0.0  # For tokenization utilities

# Cost Management and Security
cryptography>=40.0.0
keyring>=24.0.0
pyjwt>=2.6.0
sqlalchemy>=2.0.0  # For usage tracking database

# Plugin System
importlib-metadata>=6.0.0
semver>=3.0.0
pluggy>=1.0.0  # Plugin management

# Template System
jinja2>=3.1.0
markupsafe>=2.1.0
pydantic>=2.0.0  # For template validation

# Log Monitoring and Analysis
pandas>=2.0.0  # For log data analysis
matplotlib>=3.7.0  # For visualization
seaborn>=0.12.0  # For advanced visualization
tabulate>=0.9.0  # For table formatting
nltk>=3.8.0  # For text analysis
scikit-learn>=1.2.0  # For anomaly detection
dateparser>=1.1.0  # For parsing various date formats

# Docker Compose Management
jsonschema>=4.17.0  # For validating Docker Compose files
graphviz>=0.20.0  # For generating visualizations
networkx>=3.0.0  # For graph operations in visualizations

# Resource Monitoring
prometheus-client>=0.16.0  # For exposing metrics in Prometheus format
statsmodels>=0.14.0  # For time series analysis and forecasting
bokeh>=3.2.0  # For interactive visualizations
influxdb-client>=1.36.0  # For time series database integration
pyarrow>=12.0.0  # For efficient data serialization
fastparquet>=2023.0.0  # For parquet file format support
schedule>=1.2.0  # For scheduling periodic tasks

# Security Module
# Note: Using mock implementations for these in the test environment
# trivy-python>=0.1.0  # For vulnerability scanning
# docker-bench-security>=1.0.0  # For security auditing
safety>=2.3.0  # For Python package vulnerability scanning
bandit>=1.7.0  # For Python code security scanning

# Backup Module
# Note: Using standard library modules for backup functionality
# tarfile is part of the Python standard library

# Update System
semver>=3.0.0
requests>=2.28.0

# Testing
responses>=0.23.0
pytest-mock>=3.10.0
